requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.1
marshmallow-sqlalchemy==0.28.1
Werkzeug==3.0.2
python-dateutil==2.9.0.post0
requests==2.31.0
google-api-python-client==2.129.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
PyJWT==2.8.0
